transId read from the request header

+++ ups/api/src/main/java/com/s3group/tmsapi/repo/ParcelResponseRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParcelResponseRepository extends PagingAndSortingRepository<ParcelResponse, Long> {
+public interface ParcelResponseRepository extends PagingAndSortingRepository<ParcelResponse, Integer> {
   Page<ParcelResponse> findAll(Pageable pageable);
 }
